backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
google-genai
python-multipart
Pillow
